requirements.txt
Flask
gunicorn
twilio
google-cloud-dialogflow
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dateutil
openai>=1.0.0